@@ -39,7 +39,7 @@
     }
   },
   "require": {
-    "php": "^7.3|^8.0",
+    "php": "^7.4|^8.0",
     "roots/acorn": "2.0.0-beta.5"
   },
   "require-dev": {
