@@ -9,7 +9,7 @@ class App extends Composer
     /**
      * List of views served by this composer.
      *
-     * @var array
+     * @var string[]
      */
     protected static $views = [
         '*',
@@ -17,10 +17,8 @@ class App extends Composer
 
     /**
      * Data to be passed to view before rendering.
-     *
-     * @return array
      */
-    public function with()
+    public function with(): array
     {
         return [
             'siteName' => $this->siteName(),
@@ -29,10 +27,8 @@ class App extends Composer
 
     /**
      * Returns the site name.
-     *
-     * @return string
      */
-    public function siteName()
+    public function siteName(): string
     {
         return get_bloginfo('name', 'display');
     }
