@@ -12,11 +12,8 @@ use function Roots\bundle;
 
 /**
  * Register the `.brand` selector to the blogname.
- *
- * @param  \WP_Customize_Manager $wp_customize
- * @return void
  */
-add_action('customize_register', function (WP_Customize_Manager $wp_customize) {
+add_action('customize_register', function (WP_Customize_Manager $wp_customize): void {
     $wp_customize->get_setting('blogname')->transport = 'postMessage';
     $wp_customize->selective_refresh->add_partial('blogname', [
         'selector' => '.brand',
@@ -28,9 +25,7 @@ add_action('customize_register', function (WP_Customize_Manager $wp_customize) {
 
 /**
  * Register the customizer assets.
- *
- * @return void
  */
-add_action('customize_preview_init', function () {
+add_action('customize_preview_init', function (): void {
     bundle('customizer')->enqueueJs(true, ['customize-preview']);
 });
