@@ -8,9 +8,7 @@ namespace App;
 
 /**
  * Add "… Continued" to the excerpt.
- *
- * @return string
  */
-add_filter('excerpt_more', function () {
+add_filter('excerpt_more', function (): string {
     return sprintf(' &hellip; <a href="%s">%s</a>', get_permalink(), __('Continued', 'sage'));
 });
